update session table (now store only user id in session), update code according to hashed password and updated session structure

+++ server/controllers/controller.js
@@ -8,19 +8,19 @@ import formidable from "formidable"
 export async function userLogin(req, res){
     try{
         let data = await getReqData(req);
-        console.log(data)
-        let user = await actions.CheckUser(data);
+
+        let user = await actions.CheckUser(data.email);
 
         let isValid = await bcrypt.compare(data?.password ?? "", user?.password ?? "");
 
-        if(!isValid){
+        if(!isValid || !user){
             res.statusCode = 404;
             res.end(JSON.stringify({status: "failed", msg: "Invalid email or password"}));
             return;
         }
 
         let sessionId = crypto.randomBytes(8).toString('hex');
-        await actions.InsertSession(sessionId, user.email, user.password);
+        await actions.InsertSession(sessionId, user.id);
 
         res.statusCode = 200;
         res.setHeader("Set-Cookie", `sessionId=${sessionId}; HttpOnly; SameSite=Lax`);
@@ -38,7 +38,6 @@ export async function userSignup(req, res){
         let fields = form.fields;
         let files = form.files;
 
-        console.log(fields.password[0]);
         let hashedPass = await bcrypt.hash(fields.password[0], 12);
 
         let reqData = {
@@ -49,9 +48,9 @@ export async function userSignup(req, res){
             profileImage: files?.profilePhoto?.[0]?.newFilename
         }
 
-        let insertUser = await actions.InsertUser(reqData);
+        let checkUser = await actions.CheckUser(reqData.email);
 
-        if(!insertUser){
+        if(checkUser){
             if(files?.profilePhoto?.[0].filepath){
                 fs.unlink(files?.profilePhoto?.[0].filepath, (err) => {});
             }
@@ -60,8 +59,10 @@ export async function userSignup(req, res){
             return;
         }
 
+        let insertUser = await actions.InsertUser(reqData); // it return user id on inserted user
+
         let sessionId = crypto.randomBytes(8).toString('hex');
-        let insterSession = await actions.InsertSession(sessionId, reqData.email, reqData.password);
+        let insterSession = await actions.InsertSession(sessionId, insertUser);
 
         res.setHeader("Set-Cookie", `sessionId=${sessionId}; path=/; HttpOnly; SameSite=Lax`);
         res.end(JSON.stringify({status: "ok"}));
@@ -80,7 +81,7 @@ export async function getUser(req, res){
             res.end(JSON.stringify({status: "failed", msg: "Session expired. Please login again."}));
             return;
         }
-
+        
         let user = await actions.GetUser(sessionId);
 
         if(!user){
@@ -126,29 +127,39 @@ export async function userUpdate(req, res){
 
 export async function userUpdateImage(req, res){
     try{
+        let sessionId = getSessionId(req);
+        if(!sessionId){
+            res.statusCode = 401;
+            res.end(JSON.stringify({status: "failed", msg: "Session expired. Please login again."}));
+            return;
+        }
+
         let form = await uploadMedia(req);
         let fields = form.fields;
         let files = form.files;
 
         let id = parseInt(fields?.id?.[0]);
         let image = files?.profileImage?.[0].newFilename ?? "";
+        let oldImage = fields?.oldImageName?.[0] ?? "";
 
         let updateImage = await actions.UpdateProfileImage({image: image, id: id});
-
         if(!updateImage){
             res.statusCode = 400;
             res.end(JSON.stringify({status: "failed", msg: "Failed to update profile image."}));
             return;
         }
 
-        let oldImage = path.join(import.meta.dirname, "../client/public/assets/profileImages/", fields?.oldImageName?.[0])
-        if(fs.existsSync(oldImage)){
-            fs.unlink(oldImage, (err) => {});
+        if(oldImage){
+            let oldImage = path.join(import.meta.dirname, "../client/public/assets/profileImages/", )
+            if(fs.existsSync(oldImage)){
+                fs.unlink(oldImage, (err) => {});
+            }
         }
 
         res.statusCode = 200;
         res.end(JSON.stringify({status: "ok", msg: "Profile Image update successfully.", image: image}));
     }catch(err){
+        console.log(err)
         res.statusCode = 500;
         res.end(JSON.stringify({status: "error", msg: "Interal server error"}));
     }
@@ -164,8 +175,20 @@ export async function userChangePassword(req, res){
         }
 
         let data = await getReqData(req);
+        let getuser = await actions.GetUser(sessionId);
+        let user = await actions.CheckUser(getuser.email);
+
+        let isValid = await bcrypt.compare(data.currPass, user?.password ?? '');
+
+        if(!isValid){
+            res.statusCode = 400;
+            res.end(JSON.stringify({status: "failed", msg: "wrong current password !"}));
+            return;
+        }
+
+        let password = await bcrypt.hash(data.newPass, 12);
             
-        let change = await actions.ChangePassword(data.currPass, data.newPass, sessionid);
+        let change = await actions.ChangePassword(password, sessionId);
 
         if(!change){
             res.statusCode = 400;
@@ -176,6 +199,7 @@ export async function userChangePassword(req, res){
         res.statusCode = 200;
         res.end(JSON.stringify({status: "ok", msg: "Password changed successfully."}));
     }catch(err){
+        console.log(err)
         res.statusCode = 500;
         res.end(JSON.stringify({status: "error", msg: "Internal server error"}));
     }
@@ -216,16 +240,17 @@ export async function userDelete(req, res){
             return;
         }
 
-        let del = await actions.DeleteAccount(sessionid);
+        let user = await actions.GetUser(sessionId);
+        let del = await actions.DeleteAccount(sessionId);
 
-        if(!del.res){
+        if(!del){
             res.statusCode = 400;
             res.end(JSON.stringify({status: "failed", msg: "Failed to delete account !"}));
             return;
         }
 
-        if(del.img){
-            let oldImage = path.join(import.meta.dirname, "../client/public/assets/profileImages/", del.img);
+        if(user.profileImage != ''){
+            let oldImage = path.join(import.meta.dirname, "../../client/public/assets/profileImages/", user.profileImage);
             if(fs.existsSync(oldImage)){
                 fs.unlink(oldImage, (err) => {});
             }
@@ -235,6 +260,7 @@ export async function userDelete(req, res){
         res.setHeader("Set-Cookie", "sessionId=; Max-Age=0; HttpOnly");
         res.end(JSON.stringify({status: "ok", msg: "Account deleted successfully."}));
     }catch(err){
+        console.log(err);
         res.statusCode = 500;
         res.end(JSON.stringify({status: "error", msg: "Internal server error"}));
     }
@@ -242,7 +268,7 @@ export async function userDelete(req, res){
 
 function uploadMedia(request){
     let form = formidable({
-        uploadDir: path.join(import.meta.dirname, "../client/public/assets/profileImages/"),
+        uploadDir: path.join(import.meta.dirname, "../../client/public/assets/profileImages/"),
         keepExtensions: true
     });
